+++ changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.3.0] - 2025-10-06
 ### Added 
 - Added RSS Feed
+- Added new subscriber button
 
 ## [2.2.10] - 2025-10-02
 ### Added